Python wrapper: about the wrapped callbacks, maintain a 'map' of subscribers
The 'callback' function is a part of the 'Subscriber' class.
Once the 'event' (or asynchronous message) is received, the 'Subscriber' object
forwards the data to the 'Handler' object.
With this approach, the 'event' processing algorithm can be defined
in a Python subclass by the user.

Each 'subscriber' class has a function that matches the C function pointer.
But this function is a static method, shared with all instances of the class.
In order to dispatch the message/data to the right instance,
we maintain a dictionary of instantiated objects.

Therfore, we have added the following internal services :
 * bool registerNewSubscriber(EventSubscriber*, id);
 * EventSubscriber* findSubscriber(id);
 * void unregisterSubscriber(id);