tls: fix order of operations to copy peer certificate
The memcpy is done before assigning the length, so the length is not
set and is therefore either 0 (so no peer certificate will be available)
or a random number (that can lead to crashes) making the feature not work.

The MR simply copies first the length that will make the memcpy work.