quality: add 'const' qualifier to the 'MMS value' getters
Some getters already have the 'const' qualifier, but not all.
To improve the quality of the code, add the 'const' qualifier for the other
getters of 'MMS value', where it was missing.

Remark:
when a pointer is returned by a getter, this pointer must be a pointer
to a constant, so as not to modify the internal data of the 'MMS value' object.
The correctness of the function calls, in libiec61850, has been checked with
the compiler option: '-Wdiscarded-qualifiers'.

Signed-off-by: Mikael Bourhis <mikael.bourhis@smile.fr>